refactor(stateappevents): carve canary/compact source + redistribute tests
Carve two more concerns out of snapshot.go so each has its own file and
matching test (consistent with the projector split):

  stateappevents_canary.go   emitFoldDigest (cross-pod fold-divergence canary)
  stateappevents_compact.go  maybeCompact (snapshot-FIRST/compact-SECOND reclaim)

Redistribute the 1262-line merged projector test (42 funcs) to mirror the
sources, with shared doubles/helpers in one stateappevents_helpers_test.go:

  canary_internal_test.go     FoldDigest* (4)
  compact_internal_test.go    Compact*/AutoCompacts/SnapshotOnly (6)
  coldstart_internal_test.go  FreshProjectorColdStarts (1)
  gap_internal_test.go        gap classification + reseed/halt (13)
  fold_internal_test.go       epoch reset, foldVerify, determinism, fuzz (6)
  projector_internal_test.go  lifecycle: rehydrate/shutdown (4)
  snapshot_internal_test.go   += SnapshotWriteNeverRegresses (8 total)

Every source file now has a matching test; all land at/under the ~300-line
guideline (helpers_test.go excepted as test infra). Same 42 test funcs, none
lost/duplicated; gofmt + go vet + go test -race green.
