+++ stateappevents_compact_internal_test.go
@@ -1,0 +1,178 @@
+package via
+
+import (
+	"context"
+	"encoding/json"
+	"net/http/httptest"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+)
+
+// Compaction reclaims the prefix a snapshot already covers, but retained records
+// MUST keep their original offsets — a resuming pod that passed offset N must
+// still resume exactly after N. Re-numbering the survivors to 1 would silently
+// re-deliver or skip events.
+func TestCompactDropsPrefixButKeepsRetainedOffsetsStable(t *testing.T) {
+	t.Parallel()
+	b := InMemory()
+	defer b.Close()
+	ctx := context.Background()
+	for i := 1; i <= 5; i++ {
+		off, err := b.Append(ctx, "k", []byte{byte(i)})
+		require.NoError(t, err)
+		require.Equal(t, Offset(i), off)
+	}
+	c, ok := b.(Compactor)
+	require.True(t, ok, "the in-memory backplane must offer compaction")
+	require.NoError(t, c.Compact(ctx, "k", 4)) // discard offsets < 4 (1,2,3)
+
+	// Head is unchanged — compaction reclaims the prefix, never the head.
+	head, _, err := b.Head(ctx, "k")
+	require.NoError(t, err)
+	require.Equal(t, Offset(5), head)
+
+	// A genesis subscriber resumes at the lowest RETAINED offset, gap-free, with
+	// offsets UNCHANGED (4, 5) — never renumbered to 1, 2.
+	sub, err := b.Subscribe(ctx, "k", 0)
+	require.NoError(t, err)
+	require.Equal(t, Offset(4), recvOffset(t, sub))
+	require.Equal(t, Offset(5), recvOffset(t, sub))
+}
+
+// A re-issued, stale, or over-large beforeOffset must never corrupt the log or
+// move the head — compaction is idempotent and clamped to committed offsets.
+func TestCompactIsIdempotentAndClampedToCommitted(t *testing.T) {
+	t.Parallel()
+	b := InMemory()
+	defer b.Close()
+	ctx := context.Background()
+	for i := 1; i <= 3; i++ {
+		_, err := b.Append(ctx, "k", []byte{byte(i)})
+		require.NoError(t, err)
+	}
+	c := b.(Compactor)
+	require.NoError(t, c.Compact(ctx, "k", 0))   // before:0 (first-snapshot floor) → no-op
+	require.NoError(t, c.Compact(ctx, "k", 2))   // drop offset 1
+	require.NoError(t, c.Compact(ctx, "k", 2))   // again → no-op
+	require.NoError(t, c.Compact(ctx, "k", 1))   // below the floor → no-op
+	require.NoError(t, c.Compact(ctx, "k", 999)) // beyond head → clamp to committed
+
+	head, _, err := b.Head(ctx, "k")
+	require.NoError(t, err)
+	require.Equal(t, Offset(3), head, "head must never move under compaction")
+}
+
+// Offsets are a monotone resume primitive: an append AFTER compaction must
+// continue from the head, never restart base-relative. A renumbered offset would
+// let a resuming pod silently re-process or skip events.
+func TestAppendAfterCompactContinuesMonotoneOffsets(t *testing.T) {
+	t.Parallel()
+	b := InMemory()
+	defer b.Close()
+	ctx := context.Background()
+	for i := 1; i <= 5; i++ {
+		_, err := b.Append(ctx, "k", []byte{byte(i)})
+		require.NoError(t, err)
+	}
+	require.NoError(t, b.(Compactor).Compact(ctx, "k", 4)) // drop 1..3, retain 4,5
+
+	off, err := b.Append(ctx, "k", []byte{6})
+	require.NoError(t, err)
+	require.Equal(t, Offset(6), off, "append after compaction must continue from the head, not restart")
+
+	sub, err := b.Subscribe(ctx, "k", 0)
+	require.NoError(t, err)
+	require.Equal(t, Offset(4), recvOffset(t, sub))
+	require.Equal(t, Offset(5), recvOffset(t, sub))
+	require.Equal(t, Offset(6), recvOffset(t, sub))
+}
+
+// Compaction is monotone: the retained floor only ever rises. A stale or
+// out-of-order beforeOffset must not move the floor backward (which would
+// "resurrect" discarded offsets) or move the head.
+func TestSequentialCompactionsAdvanceMonotonically(t *testing.T) {
+	t.Parallel()
+	b := InMemory()
+	defer b.Close()
+	ctx := context.Background()
+	for i := 1; i <= 10; i++ {
+		_, err := b.Append(ctx, "k", []byte{byte(i)})
+		require.NoError(t, err)
+	}
+	c := b.(Compactor)
+	require.NoError(t, c.Compact(ctx, "k", 4))
+	require.Equal(t, Offset(4), lowestRetainedOffset(t, b, "k"))
+	require.NoError(t, c.Compact(ctx, "k", 7))
+	require.Equal(t, Offset(7), lowestRetainedOffset(t, b, "k"))
+	require.NoError(t, c.Compact(ctx, "k", 5)) // backward → no-op
+	require.Equal(t, Offset(7), lowestRetainedOffset(t, b, "k"), "the floor must never move backward")
+
+	head, _, err := b.Head(ctx, "k")
+	require.NoError(t, err)
+	require.Equal(t, Offset(10), head, "head is invariant under compaction")
+}
+
+// The projector reclaims storage automatically, but compaction must trail a
+// DURABLE snapshot (snapshot-FIRST, compact-SECOND) — it must never discard at
+// or beyond the covered offset, or a cold start could not resume.
+func TestProjectorAutoCompactsTrailingTheDurableSnapshot(t *testing.T) {
+	t.Parallel()
+	var server *httptest.Server
+	app := New(WithTestServer(&server), WithSnapshotInterval(1))
+	defer server.Close()
+	ctx := context.Background()
+
+	var h StateAppEvents[envEv, []int]
+	h.bindWireKey("k")
+	h.bindApp(app)
+	for i := 1; i <= 5; i++ {
+		_, err := app.backplane.Append(ctx, "k", goodEnv(t, envEv{N: i}))
+		require.NoError(t, err)
+	}
+
+	// Compaction ran: the lowest retained offset rose above genesis.
+	require.Eventually(t, func() bool {
+		return lowestRetainedOffset(t, app.backplane, "k") > 1
+	}, 2*time.Second, 10*time.Millisecond, "the projector must compact the prefix")
+
+	head, _, err := app.backplane.Head(ctx, "k")
+	require.NoError(t, err)
+	require.Equal(t, Offset(5), head)
+
+	// The durable snapshot's covered offset is never compacted away (the floor
+	// lags one generation), so a cold start can always resume from it.
+	data, _, ok, _ := app.backplane.LoadSnapshot(ctx, snapKey("k"))
+	require.True(t, ok)
+	var cp checkpoint
+	require.NoError(t, json.Unmarshal(data, &cp))
+	require.LessOrEqual(t, lowestRetainedOffset(t, app.backplane, "k"), cp.CoveredOffset,
+		"compaction must never pass the durable snapshot's covered offset")
+}
+
+// A backend that declines Compactor must still snapshot — the runtime falls back
+// to snapshot-only and never panics or wedges, leaving the prefix intact.
+func TestSnapshotOnlyWhenBackendDeclinesCompaction(t *testing.T) {
+	t.Parallel()
+	var server *httptest.Server
+	bp := nonCompactingBackplane{InMemory()}
+	app := New(WithTestServer(&server), WithBackplane(bp), WithSnapshotInterval(1))
+	defer server.Close()
+	ctx := context.Background()
+
+	var h StateAppEvents[envEv, []int]
+	h.bindWireKey("k")
+	h.bindApp(app)
+	for i := 1; i <= 5; i++ {
+		_, err := app.backplane.Append(ctx, "k", goodEnv(t, envEv{N: i}))
+		require.NoError(t, err)
+	}
+
+	require.Eventually(t, func() bool {
+		_, _, ok, _ := app.backplane.LoadSnapshot(ctx, snapKey("k"))
+		return ok
+	}, 2*time.Second, 10*time.Millisecond, "snapshot must still be written")
+	require.Equal(t, Offset(1), lowestRetainedOffset(t, app.backplane, "k"),
+		"a backend without Compactor leaves the prefix intact")
+}
