+++ stateappevents_snapshot.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"hash/fnv"
-	"strconv"
 )
 
 // checkpoint is the durable fold snapshot for a StateAppEvents key: the
@@ -113,79 +111,8 @@ func (a *App) writeSnapshot(ls *logState, key string) {
 	}
 }
 
-// maybeCompact reclaims the log prefix a DURABLE snapshot now covers. Called
-// only on a successful snapshot CAS (snapshot-FIRST, compact-SECOND). The floor
-// LAGS one snapshot generation — Compact(before:prevSnapOffset) discards only
-// what the PREVIOUS snapshot already covered, so the current snapshot's offset
-// is never truncated (cold start always resumes) and ≥1 generation of tail
-// events survives for any in-flight subscriber. A backend that declines
-// Compactor runs snapshot-only. Called serially from the single projector
-// goroutine, so prevSnapOffset needs no extra synchronization beyond ls.mu.
-func (a *App) maybeCompact(ls *logState, key string, covered Offset) {
-	c, ok := a.backplane.(Compactor)
-	if !ok {
-		return // backend declines compaction → snapshot-only
-	}
-	ls.mu.RLock()
-	floor := ls.prevSnapOffset
-	diverged := ls.diverged
-	ls.mu.RUnlock()
-	if diverged {
-		// WithFoldVerify proved this key's fold non-deterministic. NEVER compact —
-		// dropping the prefix would crystallize the bad projection into durable
-		// genesis with no path back. Keep the full log so a fixed (deterministic)
-		// build can re-fold from scratch.
-		return
-	}
-	// Never discard an event a registered side-effect consumer has not yet
-	// processed: clamp the floor to the slowest consumer's committed offset
-	// (council line 272 / T-DX-6). A consumer at offset 0 pins it at genesis.
-	if cmin, ok := a.minConsumerOffset(key); ok && cmin < floor {
-		floor = cmin
-	}
-	_ = c.Compact(context.Background(), key, floor) // best-effort; a failure just defers reclamation
-	ls.mu.Lock()
-	ls.prevSnapOffset = covered
-	ls.mu.Unlock()
-}
-
 func (a *App) setSnapRev(ls *logState, rev Rev) {
 	ls.mu.Lock()
 	ls.snapRev = rev
 	ls.mu.Unlock()
-}
-
-// emitFoldDigest publishes the cheap, unconditional fold-divergence canary
-// (council T1-SRE-7): after every advancing fold a pod emits its applied offset
-// and a digest of the resulting projection, both gauged by key. Two pods that
-// have folded a key to the same offset MUST report the same digest; a persistent
-// (key, offset)-matched digest MISMATCH across pods is fold non-determinism
-// caught before it corrupts a durable snapshot. The digest reuses the snapshot
-// codec (encodeSnap), so a key with no codec simply skips the canary — never
-// panics. Best-effort: a transient encode error just skips this sample.
-//
-// Cost: this re-encodes the FULL projection on EVERY advancing fold (O(state)
-// per event), unlike maybeSnapshot which amortizes its encode over an interval.
-// "Cheap" is relative to the cross-pod corruption it catches, not to the fold
-// itself — for a large projection on a hot key the per-event encode is real. It
-// reuses the encode the projector already performs at snapshot time, so it adds
-// no new serialization machinery, only frequency.
-func (a *App) emitFoldDigest(ls *logState, key string) {
-	ls.mu.RLock()
-	proj := ls.projection
-	off := ls.cursor
-	encode := ls.encodeSnap
-	ls.mu.RUnlock()
-	if encode == nil {
-		return
-	}
-	b, err := encode(proj)
-	if err != nil {
-		return
-	}
-	h := fnv.New32a()
-	_, _ = h.Write(b)
-	m := a.metricsOrNoop()
-	m.Gauge("via.fold.offset", float64(off), "key", key)
-	m.Gauge("via.fold.digest", float64(h.Sum32()), "key", key, "offset", strconv.FormatUint(uint64(off), 10))
 }
